Refactoring and redesign.
