+++ src/main/java/su/svn/showcase/dao/jpa/AbstractRecordDaoJpa.java
@@ -1,5 +1,5 @@
 /*
- * This file was last modified at 2020.02.13 20:38 by Victor N. Skurikhin.
+ * This file was last modified at 2020.03.04 18:20 by Victor N. Skurikhin.
  * This is free and unencumbered software released into the public domain.
  * For more information, please refer to <http://unlicense.org>
  * AbstractRecordDaoJpa.java
@@ -9,7 +9,6 @@
 package su.svn.showcase.dao.jpa;
 
 import su.svn.showcase.domain.Record;
-import su.svn.showcase.exceptions.ErrorCase;
 
 import javax.persistence.*;
 import java.time.LocalDate;
