+++ src/main/java/su/svn/showcase/dao/jpa/TagDaoEjb.java
@@ -1,8 +1,8 @@
 /*
- * This file was last modified at 2020.03.03 22:49 by Victor N. Skurikhin.
+ * This file was last modified at 2020.03.04 18:20 by Victor N. Skurikhin.
  * This is free and unencumbered software released into the public domain.
  * For more information, please refer to <http://unlicense.org>
- * RoleDaoJpa.java
+ * TagDaoEjb.java
  * $Id$
  */
 
@@ -10,26 +10,29 @@ package su.svn.showcase.dao.jpa;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import su.svn.showcase.dao.RoleDao;
-import su.svn.showcase.domain.Role;
+import su.svn.showcase.dao.TagDao;
+import su.svn.showcase.domain.Tag;
+import su.svn.showcase.utils.CollectionUtil;
 
 import javax.ejb.Stateless;
 import javax.persistence.*;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 import static su.svn.shared.Constants.Db.PERSISTENCE_UNIT_NAME;
 
 /**
- * The Role DAO implementation.
+ * The Tag DAO implementation.
  *
  * @author Victor N. Skurikhin
  */
 @Stateless
-public class RoleDaoJpa extends AbstractDaoJpa<UUID, Role> implements RoleDao {
+public class TagDaoEjb extends AbstractDaoJpa<String, Tag> implements TagDao {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(RoleDaoJpa.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(TagDaoEjb.class);
+
+    private static final Pattern validating = Pattern.compile("[\\w\\sа-яА-Я]+");
 
     @PersistenceContext(unitName = PERSISTENCE_UNIT_NAME)
     private EntityManager entityManager;
@@ -42,7 +45,7 @@ public class RoleDaoJpa extends AbstractDaoJpa<UUID, Role> implements RoleDao {
      *         is null
      */
     @Override
-    public Optional<Role> findById(UUID id) {
+    public Optional<Tag> findById(String id) {
         return jpaFindById(id);
     }
 
@@ -67,69 +70,85 @@ public class RoleDaoJpa extends AbstractDaoJpa<UUID, Role> implements RoleDao {
      *         is rolled back
      */
     @Override
-    public Optional<Role> findWhereRole(String role) {
-        return jpaFindWhereField(Role.FIND_WHERE_ROLE, "role", role);
+    public Optional<Tag> findWhereTag(String tag) {
+        return jpaFindWhereField(Tag.FIND_WHERE_TAG, "tag", tag);
     }
 
-    /**
-     * {@inheritDoc }
-     */
-    @Override
-    public List<Role> findAll() {
-        return abstractDaoFindAll(Role.FIND_ALL);
-    }
-
-    @Override
-    public List<Role> findAllOrderByRoleAsc() {
-        return abstractDaoFindAll(Role.FIND_ALL_ORDER_BY_ROLE_ASC);
-    }
-
-    @Override
-    public List<Role> findAllOrderByRoleDesc() {
-        return abstractDaoFindAll(Role.FIND_ALL_ORDER_BY_ROLE_DESC);
-    }
-
-    /**
-     * TODO
-     *
-     * {@inheritDoc }
-     */
-    @Override
-    public List<Role> findAllByIdIn(Iterable<UUID> ids) {
-        return abstractDaoFindAllWhereIn(Role.FIND_ALL_WHERE_ID_IN, "ids", ids);
-    }
-
-    @Override
-    public List<Role> findAllWhereRole(String role) {
-        return abstractDaoFindAllWhereField(Role.FIND_WHERE_ROLE, "role", role);
-    }
-
-    @Override
-    public List<Role> findAllByRoleIn(Iterable<String> roles) {
-        return abstractDaoFindAllWhereIn(Role.FIND_ALL_WHERE_ROLE_IN, "roles", roles);
-    }
-
-    @Override
-    public List<Role> range(int start, int size) {
-        return jpaRange(Role.FIND_ALL, start, size);
-    }
-
-    @Override
-    public List<Role> rangeOrderByRoleAsc(int start, int size) {
-        return jpaRange(Role.FIND_ALL_ORDER_BY_ROLE_ASC, start, size);
-    }
-
-    @Override
-    public List<Role> rangeOrderByRoleDesc(int start, int size) {
-        return jpaRange(Role.FIND_ALL_ORDER_BY_ROLE_DESC, start, size);
-    }
-
-    /**
-     * {@inheritDoc }
-     */
     @Override
     public long count() {
         return abstractCount();
+    }
+
+    /**
+     * {@inheritDoc }
+     */
+    @Override
+    public List<Tag> findAll() {
+        return abstractDaoFindAll(Tag.FIND_ALL);
+    }
+
+    /**
+     * {@inheritDoc }
+     */
+    @Override
+    public List<Tag> findAllOrderByTagAsc() {
+        return abstractDaoFindAll(Tag.FIND_ALL_ORDER_BY_TAG_ASC);
+    }
+
+    /**
+     * {@inheritDoc }
+     */
+    @Override
+    public List<Tag> findAllOrderByTagDesc() {
+        return abstractDaoFindAll(Tag.FIND_ALL_ORDER_BY_TAG_DESC);
+    }
+
+    /**
+     * {@inheritDoc }
+     */
+    @Override
+    public List<Tag> findAllWhereTag(String tag) {
+        return abstractDaoFindAllWhereField(Tag.FIND_WHERE_TAG, "tag", tag);
+    }
+
+    /**
+     * {@inheritDoc }
+     */
+    @Override
+    public List<Tag> findAllByIdIn(Iterable<String> ids) {
+        return abstractDaoFindAllWhereIn(Tag.FIND_ALL_WHERE_ID_IN, "ids", ids);
+    }
+
+    /**
+     * {@inheritDoc }
+     */
+    @Override
+    public List<Tag> findAllByTagIn(Iterable<String> tags) {
+        return abstractDaoFindAllWhereIn(Tag.FIND_ALL_WHERE_TAG_IN, "tags", tags);
+    }
+
+    /**
+     * {@inheritDoc }
+     */
+    @Override
+    public List<Tag> range(int start, int size) {
+        return jpaRange(Tag.FIND_ALL, start, size);
+    }
+
+    /**
+     * {@inheritDoc }
+     */
+    @Override
+    public List<Tag> rangeOrderByTagAsc(int start, int size) {
+        return jpaRange(Tag.FIND_ALL_ORDER_BY_TAG_ASC, start, size);
+    }
+
+    /**
+     * {@inheritDoc }
+     */
+    @Override
+    public List<Tag> rangeOrderByTagDesc(int start, int size) {
+        return jpaRange(Tag.FIND_ALL_ORDER_BY_TAG_DESC, start, size);
     }
 
     /**
@@ -144,7 +163,7 @@ public class RoleDaoJpa extends AbstractDaoJpa<UUID, Role> implements RoleDao {
      * @throws PersistenceException if the flush fails
      */
     @Override
-    public Role save(Role entity) {
+    public Tag save(Tag entity) {
         return jpaDaoSave(entity);
     }
 
@@ -152,7 +171,7 @@ public class RoleDaoJpa extends AbstractDaoJpa<UUID, Role> implements RoleDao {
      * {@inheritDoc }
      */
     @Override
-    public Iterable<Role> saveAll(Iterable<Role> entities) {
+    public Iterable<Tag> saveAll(Iterable<Tag> entities) {
         return abstractDaoSaveAll(entities);
     }
 
@@ -160,7 +179,7 @@ public class RoleDaoJpa extends AbstractDaoJpa<UUID, Role> implements RoleDao {
      * {@inheritDoc }
      */
     @Override
-    public void delete(UUID id) {
+    public void delete(String id) {
         abstractDaoDelete(id);
     }
 
@@ -168,8 +187,32 @@ public class RoleDaoJpa extends AbstractDaoJpa<UUID, Role> implements RoleDao {
      * {@inheritDoc }
      */
     @Override
-    public void deleteAll(Iterable<Role> entities) {
+    public void deleteAll(Iterable<Tag> entities) {
         abstractDaoDeleteAll(entities);
+    }
+
+    private boolean isValidListOfTags(Iterable<String> tags) {
+        for (String tag : tags) {
+            if ( ! validating.matcher(tag).matches()) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    /**
+     * {@inheritDoc }
+     */
+    @Override
+    public List<String> outerSection(Iterable<String> tags) {
+        if ( ! isValidListOfTags(tags)) {
+            throw new RuntimeException("Isn't valid tag in list: " + tags);
+        }
+        String values = CollectionUtil.iterableToStream(tags)
+                .map(s -> "('" + s + "')")
+                .collect(Collectors.joining(","));
+        String sql = String.format(Tag.OUTER_SECTION, values);
+        return abstractDaoNativeResultList(sql, String.class);
     }
 
     /**
@@ -180,9 +223,6 @@ public class RoleDaoJpa extends AbstractDaoJpa<UUID, Role> implements RoleDao {
         return this.entityManager;
     }
 
-    /**
-     * {@inheritDoc }
-     */
     @Override
     Logger getLogger() {
         return LOGGER;
@@ -192,8 +232,8 @@ public class RoleDaoJpa extends AbstractDaoJpa<UUID, Role> implements RoleDao {
      * {@inheritDoc }
      */
     @Override
-    public Class<Role> getEClass() {
-        return Role.class;
+    public Class<Tag> getEClass() {
+        return Tag.class;
     }
 }
 //EOF
