+++ src/test/java/su/svn/showcase/services/impl/UserLoginDaoEjbTest.java
@@ -1,8 +1,8 @@
 /*
- * This file was last modified at 2020.02.18 11:26 by Victor N. Skurikhin.
+ * This file was last modified at 2020.02.13 21:57 by Victor N. Skurikhin.
  * This is free and unencumbered software released into the public domain.
  * For more information, please refer to <http://unlicense.org>
- * RoleDaoJpaTest.java
+ * UserLoginDaoJpaTest.java
  * $Id$
  */
 
@@ -12,14 +12,14 @@ import org.jboss.weld.junit5.WeldInitiator;
 import org.jboss.weld.junit5.WeldJunit5Extension;
 import org.jboss.weld.junit5.WeldSetup;
 import org.jboss.weld.junit5.auto.AddPackages;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import su.svn.showcase.dao.RoleDao;
-import su.svn.showcase.dao.jpa.RoleDaoJpa;
-import su.svn.showcase.domain.Role;
-import su.svn.showcase.domain.TestData;
+import su.svn.showcase.dao.UserLoginDao;
+import su.svn.showcase.dao.jpa.UserLoginDaoEjb;
+import su.svn.showcase.domain.UserLogin;
 import su.svn.showcase.services.impl.support.EntityManagerFactoryProducer;
 import su.svn.showcase.services.impl.support.EntityManagerProducer;
 import su.svn.showcase.services.impl.support.JtaEnvironment;
@@ -34,17 +34,20 @@ import javax.persistence.Persistence;
 import javax.transaction.NotSupportedException;
 import javax.transaction.SystemException;
 import javax.transaction.UserTransaction;
-import java.util.*;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.function.Function;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.mock;
+import static su.svn.showcase.domain.TestData.clean;
+import static su.svn.showcase.domain.TestData.cloneUserLogin1;
 import static su.svn.showcase.services.impl.support.EntityManagerFactoryProducer.configure;
 
-@DisplayName("A RoleDaoJpaTest unit test cases")
-@AddPackages(value = {RoleDaoJpa.class})
+@DisplayName("A UserLoginDaoTest unit test cases")
+@AddPackages(value = {UserLoginDao.class})
 @ExtendWith({JtaEnvironment.class, WeldJunit5Extension.class})
-class RoleDaoJpaTest {
+class UserLoginDaoEjbTest {
 
     @Inject
     private BeanManager beanManager;
@@ -54,7 +57,7 @@ class RoleDaoJpaTest {
     @WeldSetup
     private
     WeldInitiator weld = WeldInitiator.from(
-            RoleDaoJpa.class,
+            UserLoginDaoEjb.class,
             EntityManagerFactoryProducer.class,
             EntityManagerProducer.class)
             .activate(RequestScoped.class)
@@ -63,12 +66,10 @@ class RoleDaoJpaTest {
             .setPersistenceUnitFactory(injectionPoint -> emf)
             .inject(this)
             .build();
-
-    private RoleDao mockDao = mock(RoleDao.class);
-
+    private UserLoginDao mockUserLoginDao = mock(UserLoginDao.class);
     private Map<String, Object> ejbMap = new HashMap<String, Object>() {{
-        put(null, mockDao);
-        put(RoleDao.class.getName(), mockDao);
+        put(null,                      mockUserLoginDao);
+        put(UserLoginDao.class.getName(), mockUserLoginDao);
     }};
 
     private Function<InjectionPoint, Object> ejbFactory() {
@@ -81,11 +82,11 @@ class RoleDaoJpaTest {
     @Inject
     private UserTransaction userTransaction;
 
-    private Role entity;
+    private UserLogin entity;
 
     @BeforeEach
     void createNew() {
-        entity = TestData.cloneRole1();
+        entity = clean(cloneUserLogin1());
     }
 
     @DisplayName("Can inject entity manager and user transaction")
@@ -95,57 +96,14 @@ class RoleDaoJpaTest {
         assertNotNull(userTransaction);
     }
 
-    @DisplayName("Test when RoleDaoJpa findById return empty")
+    @DisplayName("Test when newsEntryDao save success")
     @Test
-    void whenRoleDao_findById_shouldBeReturnEmptyOptional() throws SystemException, NotSupportedException {
+    void whenUserLoginDao_save_success() throws SystemException, NotSupportedException {
         userTransaction.begin();
-        RoleDao dao = weld.select(RoleDaoJpa.class).get();
-        Optional<Role> test = dao.findById(UUID.randomUUID());
-        assertNotNull(test);
-        assertFalse(test.isPresent());
-        userTransaction.rollback();
-    }
-
-    @DisplayName("Test when RoleDaoJpa save success")
-    @Test
-    void whenRoleDao_findAll_shouldBeReturnEmptyList() throws SystemException, NotSupportedException {
-        userTransaction.begin();
-        RoleDao dao = weld.select(RoleDaoJpa.class).get();
-        List<Role> testList = dao.findAll();
-        assertNotNull(testList);
-        assertTrue(testList.isEmpty());
-        userTransaction.rollback();
-    }
-
-    @DisplayName("Test when RoleDaoJpa save is success")
-    @Test
-    void whenRoleDao_save_success() throws SystemException, NotSupportedException {
-        userTransaction.begin();
-        RoleDao dao = weld.select(RoleDaoJpa.class).get();
-        Optional<Role> test = dao.findById(UUID.randomUUID());
-        assertNotNull(test);
-        assertFalse(test.isPresent());
-        userTransaction.rollback();
-    }
-
-    @DisplayName("Test when RoleDaoJpa save of set is success")
-    @Test
-    void whenRoleDao_save_iterable_success() throws SystemException, NotSupportedException {
-        userTransaction.begin();
-        RoleDao dao = weld.select(RoleDaoJpa.class).get();
-        List<Role> testList = new ArrayList<Role>() {{ add(entity); }};
-        Iterable<Role> result = dao.saveAll(testList);
-        assertNotNull(result);
-        assertEquals(testList, result);
-        userTransaction.rollback();
-    }
-
-    @DisplayName("Test when RoleDaoJpa delete failed")
-    @Test
-    void whenRoleDao_delete_shouldBeReturnFalse() throws SystemException, NotSupportedException {
-        userTransaction.begin();
-        RoleDao dao = weld.select(RoleDaoJpa.class).get();
-        dao.delete(UUID.randomUUID());
+        UserLoginDao dao = weld.select(UserLoginDaoEjb.class).get();
+        UserLogin test = dao.save(entity);
+        Assertions.assertNotNull(test);
+        Assertions.assertEquals(entity, test);
         userTransaction.rollback();
     }
 }
